Vars/PyPdf/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
ci-info==0.2.0
click==8.0.4
configobj==5.0.6
configparser==5.2.0
etelemetry==0.3.0
filelock==3.6.0
fitz==0.0.1.dev2
future==0.18.2
httplib2==0.20.4
idna==3.3
isodate==0.6.1
lxml==4.8.0
mysql-connector-python==8.0.28
mysql-connector-repackaged==0.3.1
networkx==2.7
nibabel==3.2.2
nipype==1.7.0
numpy==1.22.2
packaging==21.3
pandas==1.4.1
pathlib==1.0.1
pkg_resources==0.0.0
protobuf==3.19.4
prov==2.0.0
pydot==1.4.2
PyMuPDF==1.19.5
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pyxnat==1.4
rdflib==6.1.1
requests==2.27.1
scipy==1.8.0
simplejson==3.17.6
six==1.16.0
traits==6.3.2
urllib3==1.26.8
